Add initial graphics interface and configurations files
Implemented a basic JFrame supporting a JPanel for graphics. This commit includes the addition of src/it/unibs/pajc/Graphics.java and src/it/unibs/pajc/PanelGraphics.java files to manage graphical window and panel respectively, and the configuration settings files for the project. These changes are intended to setup the initial graphic interface for our application.