Add anti-aliasing and draw circle in PanelGraphics
Included anti-aliasing rendering hint in paintComponent of PanelGraphics for an enhanced, smoother visual output. Added a new method, fillCircle, to fill a circle in the panel. This method is used to draw a blue circle at the center of the JPanel which provides an additional graphic element to the panel. Furthermore, a module name has been aligned in the .idea directory to maintain a logical naming structure across all modules.